decomp/nerfvq_nfr3/requirements.txt
absl_py==0.15.0
dm_sonnet==2.0.0
google3==1.0.0
ipython==7.16.1
Imath
jsonpickle
matplotlib==3.3.4
nbformat
numpy==1.19.2
OpenEXR==1.3.9
Pillow==8.3.1
PyQt5
PyYAML==6.0.1
scikit_learn==0.20.1
scipy==1.4.1
scikit_image==0.13.0
sonnet==0.1.6
tensorflow_gpu==2.4.1
tensorflow_probability==0.12.1
torch==1.9.1
torchvision==0.10.1
torchaudio==0.9.1
tqdm==4.63.0
trimesh==3.15.8
